refactor player, add domain and application layers

+++ Assets/_Project/_Common/Scripts/Signals/AddThrowForcePercentSignal.cs
@@ -2,6 +2,6 @@
 {
     public struct AddThrowForcePercentSignal
     {
-        public float percent;
+        public int percent;
     }
 }
